+++ src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 
-import PageLogin from './pages/PageLogin';
-import PageReg from './pages/PageReg';
+import PageMain from './pages/PageMain';
 import PageMap from './pages/PageMap';
 import PageProfile from './pages/PageProfile';
 import { AuthProvider } from './AuthContext';
@@ -14,11 +13,10 @@ const App = () => {
 		<AuthProvider>
 			<div className='App'>
 				<Routes>
-					<Route path="/" element={<PageLogin />} exact></Route>
-					<Route path="/reg" element={<PageReg />}></Route>
+					<Route path="/" element={<PageMain />} exact></Route>
 					<Route path="/map" element={<PageMap />}></Route>
 					<Route path="/profile" element={<PageProfile />}></Route>
-					<Route path="*" element={<PageLogin />}></Route>
+					<Route path="*" element={<PageMain />}></Route>
 				</Routes>
 			</div>
 		</AuthProvider>
